cleanup: move dealer from toolbar to settings
