+++ src/features/settings/Dealer.tsx
