+++ src/features/settings/index.ts
@@ -1,2 +1,3 @@
 export * from './Settings';
 export * from './Statistics';
+export * from './Dealer';
