+++ src/features/toolbar/index.ts
@@ -3,5 +3,4 @@ export * from './Toolbar';
 export * from './RestartPopover';
 export * from './SettingsPopover';
 export * from './SettingSwitch';
-export * from './Dealer';
 export * from './StatisticsDisplay';
